Update Student_Mngmnt_Group_Project.java
Changed the format of the reportStudent method, so that it displays a student of a class's information all on one line, allowing Case 4 to display all students in a more streamlined manner, and allowing it to display the next student's info immediately after displaying the previous student.
Changed Case 4 to compensate for previous change. Student Names, IDs, Class Names, and Grades all display in a uniform column now.