Here's a better version of the original code
'Better' may be subjective.
I got a wild desire and fixed little problems I had with the original project. Hindsight is a hell of a drug.
Should I be doing finals? Yes. Is this a total waste of my time? Also yes.
But it is at least 28% better than the original, and is wayyy easier to read.